Some refactoring of classes placement #48
