+++ src/schemas/network_schemas.py
@@ -1,24 +1,12 @@
 """ Schemas for networks"""
 
 from ipaddress import IPv4Network
-from json import JSONEncoder
 from typing import Iterable
 
 from .base_input_schema import BaseInputSchema
 
 IPv4NetworkList = list[IPv4Network]
 IPv4NetworkIterable = Iterable[IPv4Network]
-
-
-class IpNetworkEncoder(JSONEncoder):
-    def __init__(self, *args, **kwargs):
-        JSONEncoder.__init__(self, *args, **kwargs)
-
-    def default(self, o):
-        if type(o) is IPv4Network:
-            return str(o)
-
-        return JSONEncoder.default(self, o)
 
 
 class AgentNetworkInfo(BaseInputSchema):
