+++ src/service/history_db_service.py
@@ -1,6 +1,6 @@
 from src.core.settings import HISTORY_USAGE_INFO
+from src.models.usage_encoders import UsageClassesEncoder
 from src.schemas.usage_schemas import AddressHistoryRecord
-from src.schemas.usage_schemas import UsageClassesEncoder
 from src.service.abstract_hkey_db_service import AbstractHkeyDBService
